Modify: NicknameForm 컴포넌트 추출
